# [Core] Use numpy to speed up padded token processing