@DarkLight1337 2024-07-15
This seems to share quite a bit in common with `vllm.utils.make_tensor_with_pad` (recently optimized by #4196). I wonder why the utils function wasn't being used in the first place?